:sparkles: Add token validation on login
This commit introduces a token validation check during the login process. The application now verifies the validity of the provided GitHub access token by making a request to the GitHub API. If the token is valid, the user is allowed to proceed to the home view. Otherwise, an error message is displayed, prompting the user to ensure the token is correct and that there is an active internet connection.